#31 feat : 모집 기능 repostiroy를 Spring data jpa로 변경
